[Decomp][Guild] Cohesion split: extract rank management to GuildRank.cpp
